firereach/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.9.2
pydantic-settings>=2.5.2
groq>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.3
duckduckgo-search>=6.3.0
python-dotenv>=1.0.1
